create dynamic path for the route
