+++ client/src/pages/Admin.js
@@ -1,20 +1,9 @@
 /* eslint-disable jsx-a11y/label-has-associated-control */
 import React from "react";
-import { useState,useEffect } from "react";
+// import { useState,useEffect } from "react";
 
-function AdminPage() {
-  const [pages, setPages] = useState([]);
+function AdminPage({pages}) {
 
-  async function fetchPagesAPI() {
-    const page = await fetch("/api/pages");
-    const response = await page.json();
-    setPages(response);
-	
-  }
- 
-  useEffect(() => {
-     fetchPagesAPI();
-  }, []);
   
 	return (
 		<div>
